Sprint 2
This sprint consists of next set of user stories. Tables created for relevant entities and relevant data was added for tables using website https://www.generatedata.com/.

Developed a more simpler Jswing user interface, so that any sort of query can be executed from the user interface for  CRUD operation on the database.