Final Sprint
This is the final sprint of the Airport Database Project. This contians the report which lists the throghout progress of this project.

This sprint majorly focusses on stored programs, indexes relevant for this Database Systems. I have created stored functions, views, triggers, events and stored procedures required for this database system. I have also worked on the indexing for all the tables and created indexes wherever required for the tables that are queried more often.